refactor completion to separate package
