+++ cmd/eksctl/main.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/kubicorn/kubicorn/pkg/logger"
 	"github.com/spf13/cobra"
+	"github.com/weaveworks/eksctl/pkg/ctl/completion"
 	"github.com/weaveworks/eksctl/pkg/ctl/create"
 	"github.com/weaveworks/eksctl/pkg/ctl/delete"
 	"github.com/weaveworks/eksctl/pkg/ctl/get"
@@ -23,56 +24,6 @@ var rootCmd = &cobra.Command{
 	},
 }
 
-var bashCompletionCmd = &cobra.Command{
-	Use:   "bash",
-	Short: "Generates bash completion scripts",
-	Long: `To load completion run
-
-. <(eksctl completion bash)
-
-To configure your bash shell to load completions for each session add to your bashrc
-
-# ~/.bashrc or ~/.profile
-. <(eksctl completion bash)
-`,
-	Run: func(cmd *cobra.Command, args []string) {
-		rootCmd.GenBashCompletion(os.Stdout)
-	},
-}
-var zshCompletionCmd = &cobra.Command{
-	Use:   "zsh",
-	Short: "Generates zsh completion scripts",
-	Long: `To load completion run
-
-. <(eksctl completion zsh)
-
-To configure your zsh shell to load completions for each session add to your zshrc
-
-# ~/.zshrc or ~/.profile
-. <(eksctl completion zsh)
-`,
-	Run: func(cmd *cobra.Command, args []string) {
-		rootCmd.GenZshCompletion(os.Stdout)
-	},
-}
-
-// completionCmd will create the `completion` commands
-func completionCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "completion",
-		Short: "Generates shell completion scripts",
-		Run: func(c *cobra.Command, _ []string) {
-			if err := c.Help(); err != nil {
-				logger.Debug("ignoring error %q", err.Error())
-			}
-		},
-	}
-
-	cmd.AddCommand(bashCompletionCmd)
-	cmd.AddCommand(zshCompletionCmd)
-
-	return cmd
-}
 func init() {
 
 	addCommands()
@@ -95,5 +46,5 @@ func addCommands() {
 	rootCmd.AddCommand(get.Command())
 	rootCmd.AddCommand(scale.Command())
 	rootCmd.AddCommand(utils.Command())
-	rootCmd.AddCommand(completionCmd())
+	rootCmd.AddCommand(completion.Command(rootCmd))
 }
